Add refactorings.
