+++ src/xcode/ENA/HealthCertificateToolkit/Tests/HealthCertificateToolkitTests/DigitalGreenCertificateAccessTests.swift
@@ -7,7 +7,7 @@ import XCTest
 import SwiftCBOR
 @testable import HealthCertificateToolkit
 
-final class HealthCertificateAccessTests: XCTestCase {
+final class DigitalGreenCertificateAccessTests: XCTestCase {
 
     func test_When_decodeSucceeds_Then_CorrectDataIsReturned() throws {
         let certificateAccess = DigitalGreenCertificateAccess()
